refactor getCustomSerializer() method
